Fix on reservation docker and stop script

+++ stopall.sh
@@ -1,2 +1,5 @@
 docker stop $(docker ps -a -q)
-docker rm $(docker ps -a -q)
+
+if [ $1 = "-d" ]; then
+	docker rm $(docker ps -a -q)
+fi
